Update inputSendMail
This code will take a picture when motion is detected then save it to the drive. We want to change it such that when it takes the picture it attaches the file to the email. Something else we were thinking is that we can change it so that the sensor will activate the cameras video and the email will send a link to the IP host url where the user can actively see what is going on. Currently working on that.